requirements.txt
anyio==3.7.1
apt-xapian-index==0.49
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.1
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
atomicwrites==1.4.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==6.0.0
blinker==1.8.2
Bootstrap-Flask==2.2.0
cachetools==5.3.3
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
colorgram.py==1.2.0
colour==0.1.5
comm==0.1.4
command-not-found==0.3
contourpy==1.1.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
discodos==1.0rc2
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
docopt==0.6.2
dominate==2.5.2
dropbox==11.26.0
easygui==0.98.1
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
fonttools==4.42.0
fqdn==1.5.1
fuse-python==1.0.2
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gpg==1.16.0
gphoto2==1.9.0
greenlet==2.0.2
gunicorn==19.7.1
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.25.1
ipython==8.14.0
iso3166==2.1.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
JACK-Client==0.5.3
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
MLB-StatsAPI==1.7.1
more-itertools==8.10.0
multidict==6.0.4
musicbrainzngs==0.7.1
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
netifaces==0.11.0
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
overrides==7.4.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.10.0
plotly==5.20.0
pluggy==0.13.0
ply==3.11
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyalsaaudio==0.8.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pyenchant==3.2.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylibacl==0.6.0
pyliblo==0.10.0
pymediainfo==5.0.3
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PySocks==1.7.1
pytest==6.2.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.18.0
python-json-logger==2.0.7
python3-discogs-client==2.3.5
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
rapid-photo-downloader==0.9.33
rdflib==6.1.1
referencing==0.30.2
reportlab==3.6.8
requests==2.24.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
scour==0.38.2
scrapingbee==2.0.1
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.16.0
Send2Trash==1.8.2
show-in-file-manager==1.1.4
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
sortedcontainers==2.1.0
soundconverter==4.0.5
soupsieve==2.3.1
SPARQLWrapper==1.8.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.2
stone==3.3.1
systemd-python==234
tabulate==0.8.9
tenacity==6.3.1
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.64.1
traitlets==5.9.0
trio==0.23.2
trio-websocket==0.11.1
twilio==7.16.0
typing_extensions==4.6.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.25.10
usb-creator==0.3.7
visitor==0.1.3
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webdavclient3==3.14.6
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
